[MIN-475] Implement CurrencyConverter
